test: refactor all tests to seperate files and seperate api and unit tests
