+++ tests/require.js
@@ -1,5 +1,6 @@
 require('dotenv').config();
 require('chai').use(require('chai-http'));
 require('chai').use(require('chai-as-promised'));
+require('chai').use(require('chai-each'));
 require('chai').should();
 require('mocha-steps');
